Allow "" as a valid Path, with some limitations
RFC6902 specifies that the "path" member of an operation is a
JSON Pointer (RFC6901), but RFC6901 clearly specifies that an empty
path refers to the root document.

https://tools.ietf.org/html/rfc6902#section-4

https://tools.ietf.org/html/rfc6901#section-5

This commit special-cases partial objects and arrays for some possible
empty path cases.

- findObject now returns the given container if path is "".
- partialDoc set, get, and remove now support empty path.
- partialArray set, get, and remove now support empty path.
- set currently only supports setting to the same type of object.

Other methods of the container interface return an error if an empty
path is given.

Tests currently only cover the valid "replace" case for partialDoc.