Remove the v5/cmd directory.
Instead, use the top-level cmd/ directory.

Point cmd/json-patch to the v5/ version of the library instead of the
top-level version. It looks like updates are being applied only to v5/
instead of both v5/ and the top-level version. We should use the most
up-to-date version of the library in the command line tool to prevent
user confusion (this should fix #150).